Added documentation and abstracted the updaters.
