+++ metron-platform/metron-common/src/main/java/org/apache/metron/common/zookeeper/ConfigurationsCache.java
@@ -17,11 +17,28 @@
  */
 package org.apache.metron.common.zookeeper;
 
-import org.apache.curator.framework.CuratorFramework;
 import org.apache.metron.common.configuration.Configurations;
 
+/**
+ * A cache for multiple Configurations object.  This cache is generally kept in
+ * sync with zookeeper changes.
+ */
 public interface ConfigurationsCache extends AutoCloseable{
+  /**
+   * Return the Configurations object given the specific type of Configurations object.
+   * @param configClass The Configurations class to return
+   * @param <T> The type of Configurations class.
+   * @return The Configurations object
+   */
   <T extends Configurations> T get(Class<T> configClass);
+
+  /**
+   * Reset the cache and reload from zookeeper.
+   */
   void reset();
+
+  /**
+   * Start the cache.
+   */
   void start();
 }
