+++ metron-platform/metron-common/src/main/java/org/apache/metron/common/zookeeper/configurations/EnrichmentUpdater.java
@@ -61,32 +61,18 @@ public class EnrichmentUpdater extends ConfigurationsUpdater<EnrichmentConfigura
   }
 
   @Override
-  public void delete(CuratorFramework client, String path, byte[] data) throws IOException {
-    String name = path.substring(path.lastIndexOf("/") + 1);
-    if (path.startsWith(ConfigurationType.ENRICHMENT.getZookeeperRoot())) {
-      LOG.debug("Deleting enrichment {} config from internal cache", name);
-      getConfigurations().delete(name);
-      reloadCallback(name, ConfigurationType.ENRICHMENT);
-    } else if (ConfigurationType.GLOBAL.getZookeeperRoot().equals(path)) {
-      LOG.debug("Deleting global config from internal cache");
-      getConfigurations().deleteGlobalConfig();
-      reloadCallback(name, ConfigurationType.GLOBAL);
-    }
+  public ConfigurationType getType() {
+    return ConfigurationType.ENRICHMENT;
   }
 
   @Override
-  public void update(CuratorFramework client, String path, byte[] data) throws IOException {
-    if (data.length != 0) {
-      String name = path.substring(path.lastIndexOf("/") + 1);
-      if (path.startsWith(ConfigurationType.ENRICHMENT.getZookeeperRoot())) {
-        LOG.debug("Updating the enrichment config: {} -> {}", name, new String(data == null?"".getBytes():data));
-        getConfigurations().updateSensorEnrichmentConfig(name, data);
-        reloadCallback(name, ConfigurationType.ENRICHMENT);
-      } else if (ConfigurationType.GLOBAL.getZookeeperRoot().equals(path)) {
-        LOG.debug("Updating the global config: {}", new String(data == null?"".getBytes():data));
-        getConfigurations().updateGlobalConfig(data);
-        reloadCallback(name, ConfigurationType.GLOBAL);
-      }
-    }
+  public void delete(String name) {
+    getConfigurations().delete(name);
   }
+
+  @Override
+  public void update(String name, byte[] data) throws IOException {
+    getConfigurations().updateSensorEnrichmentConfig(name, data);
+  }
+
 }
