+++ metron-platform/metron-common/src/main/java/org/apache/metron/common/zookeeper/configurations/ParserUpdater.java
@@ -57,31 +57,18 @@ public class ParserUpdater extends ConfigurationsUpdater<ParserConfigurations> {
   }
 
   @Override
-  public void delete(CuratorFramework client, String path, byte[] data) throws IOException {
-    String name = path.substring(path.lastIndexOf("/") + 1);
-    if (path.startsWith(ConfigurationType.PARSER.getZookeeperRoot())) {
-      LOG.debug("Deleting parser {} config from internal cache", name);
-      getConfigurations().delete(name);
-      reloadCallback(name, ConfigurationType.PARSER);
-    } else if (ConfigurationType.GLOBAL.getZookeeperRoot().equals(path)) {
-      LOG.debug("Deleting global config from internal cache");
-      getConfigurations().deleteGlobalConfig();
-      reloadCallback(name, ConfigurationType.GLOBAL);
-    }
+  public ConfigurationType getType() {
+    return ConfigurationType.PARSER;
   }
+
   @Override
-  public void update(CuratorFramework client, String path, byte[] data) throws IOException {
-    if (data.length != 0) {
-      String name = path.substring(path.lastIndexOf("/") + 1);
-      if (path.startsWith(ConfigurationType.PARSER.getZookeeperRoot())) {
-        LOG.debug("Updating the parser config: {} -> {}", name, new String(data == null?"".getBytes():data));
-        getConfigurations().updateSensorParserConfig(name, data);
-        reloadCallback(name, ConfigurationType.PARSER);
-      } else if (ConfigurationType.GLOBAL.getZookeeperRoot().equals(path)) {
-        LOG.debug("Updating the global config: {}", new String(data == null?"".getBytes():data));
-        getConfigurations().updateGlobalConfig(data);
-        reloadCallback(name, ConfigurationType.GLOBAL);
-      }
-    }
+  public void delete(String name) {
+    getConfigurations().delete(name);
   }
+
+  @Override
+  public void update(String name, byte[] data) throws IOException {
+    getConfigurations().updateSensorParserConfig(name, data);
+  }
+
 }
